Added pages, buttons, bankcard
