+++ package.json
@@ -13,7 +13,8 @@
     "@mdi/js": "^7.4.47",
     "@mdi/react": "^1.6.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
