+++ src/App.jsx
@@ -1,9 +1,17 @@
-import Card from "./components/Card";
+import { AddCard } from "./pages/addcard/AddCard";
+import { Home } from "./pages/home/Home";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 function App() {
   return (
     <>
-      <Card />
+      {/* <Home /> */}
+      <BrowserRouter>
+        <Routes>
+          <Route path="AddCard" element={<AddCard />}></Route>
+          <Route path="/" element={<Home />}></Route>
+        </Routes>
+      </BrowserRouter>
     </>
   );
 }
