+++ src/components/button/style.css
@@ -1,0 +1,25 @@
+@import url('https://fonts.googleapis.com/css2?family=PT+Mono&display=swap');
+
+.button-container{
+    display: flex;
+    justify-content: center;
+}
+
+.button{
+    width: 382px;
+    height: 80px;
+    border-radius: 8px;
+    border: 1px 0px 0px 0px;
+    font-size: 22px;
+    font-weight: 700;
+    font-family: "PT Mono", monospace;
+}
+
+.button--home{
+    background-color: white;
+}
+
+.button--add-card{
+    background-color: black;
+    color: white;
+}
