+++ src/components/pagetitle/PageTitle.jsx
@@ -1,11 +1,13 @@
 import "./style.css";
 
-export function PageTitle({ title }) {
+export function PageTitle({ title, secondTitle }) {
   const titlePage = title;
+  const underTitle = secondTitle;
 
   return (
     <header className="page-header">
       <h1 className="header-title">{titlePage}</h1>
+      <h5 className="under-title">{underTitle}</h5>
     </header>
   );
 }
