+++ src/components/pagetitle/style.css
@@ -3,7 +3,9 @@
 .page-header{
     display: flex;
     justify-content: center;
-    height: 100px;
+    align-items: center;
+    flex-direction: column;
+    
     margin-top: 3rem;
 }
 
@@ -12,6 +14,14 @@
     font-weight: 700;
     width: 13rem;
     text-align: center;
+ 
+
 }
 
+
+.under-title{
+    margin: 5px 0;
+    color: rgba(34, 34, 34, 0.6);
+    font-size: 12px;
+}
